refactor(server): split runtime config resolver types
