+++ server/proliferate/server/cloud/runtime_config/service.py
@@ -48,13 +48,13 @@ from proliferate.server.cloud.runtime_config.domain.manifest import (
     CompiledRuntimeConfigManifest,
     compile_runtime_config_manifest,
 )
-from proliferate.server.cloud.runtime_config.domain.resolver import (
+from proliferate.server.cloud.runtime_config.domain.resolver import resolve_runtime_config
+from proliferate.server.cloud.runtime_config.domain.types import (
     McpConnectionSnapshot,
     PluginConfiguredItemSnapshot,
     ResolverInput,
     SandboxProfileResolverSnapshot,
     SkillConfiguredItemSnapshot,
-    resolve_runtime_config,
 )
 from proliferate.server.cloud.runtime_config.models import (
     DesktopRuntimeConfigApplyResponse,
